rmd-22827 Refactor BeginScan and improve tests
