+++ README.adoc
@@ -83,6 +83,9 @@ Tested with the following XML file versions:
 * detailedreport: report_format_version="1.5"
 * buildlist: buildlist_version="1.3"
 * filelist: filelist_version="1.1"
+
+* prescanresults: prescanresults_version="1.4"
+
 * sandboxlist: sandboxlist_version="1.0"
 
 * sandboxinfo: sandboxinfo_version="1.2"
