+++ src/main/groovy/com/calgaryscientific/gradle/VeracodePreScanResults.groovy
@@ -35,4 +35,76 @@ class VeracodePreScanResults {
             printf "id=%s name=\"%s\" status=\"%s\"\n", XMLIO.getNodeAttributes(module, 'id', 'name', 'status')
         }
     }
+
+    /**
+     * Given Veracode's GetPreScanResults xml and a whitelist list of modules,
+     * it will return the module IDs for the non fatal whitelisted modules.
+     *
+     * @param xml
+     * @param whitelist
+     * @return
+     */
+    static Set<String> extractWhitelistModuleIds(Node xml, Set<String> whitelist) {
+        List<Node> nonFatalModules = filterOutFatalModules(XMLIO.getNodeList(xml, 'module'))
+        List<Node> whitelistModules = getWhitelistModules(nonFatalModules, whitelist)
+        printMissingWhitelistModules(whitelist, whitelistModules)
+        return whitelistModules.collect { module ->
+            module.attribute('id') as String
+        }.toSet()
+    }
+
+    private static void printMissingWhitelistModules(Set<String> whitelist, List<Node> whitelistModules) {
+        Set<String> missingWhitelistModules = getMissingWhitelistModules(whitelist, whitelistModules)
+        if (missingWhitelistModules.size() > 0) {
+            // TODO: Look into logging levels. The whole plugin is using print statements.
+            printf "WARNING: Missing whitelist modules: ${missingWhitelistModules}\n"
+        }
+    }
+
+    /**
+     * Given a Veracode moduleList List<Node> (xml), it will filter out modules that are not fatal.
+     *
+     * @param moduleList
+     * @return Map{ name: [id, status] }
+     */
+    private static List<Node> filterOutFatalModules(List<Node> moduleList) {
+        moduleList.findAll { module ->
+            !(module.attribute('status') as String).startsWith('(Fatal)')
+        }
+    }
+
+    /**
+     * Given a list of modules, it will return the subset of modules that are found in the whitelist.
+     * @param modules
+     * @param whitelist
+     * @return Map{ name: [id, status] }
+     */
+    private
+    static List<Node> getWhitelistModules(List<Node> modules, Set<String> whitelist) {
+        modules.findAll { module ->
+            if (module && whitelist && whitelist.contains(module.attribute('name') as String)) {
+                printf "Selecting module: %s: %s - %s\n", XMLIO.getNodeAttributes(module, 'id', 'name', 'status')
+                return true
+            }
+        }
+    }
+
+    /**
+     * Given a whitelist and a list of whitelistModules,
+     * it will return the whitelist entries that aren't part of the whitelist modules.
+     * @param whitelist
+     * @param whitelistModules
+     * @return
+     */
+    private
+    static Set<String> getMissingWhitelistModules(Set<String> whitelist, List<Node> whitelistModules) {
+        Set<String> missingWhitelistModules = []
+        if (whitelist && whitelist.size() != whitelistModules.size()) {
+            Set<String> moduleNames = whitelistModules.collect { module ->
+                module.attribute('name') as String
+            }.toSet()
+            missingWhitelistModules = whitelist - moduleNames
+        }
+        missingWhitelistModules
+    }
 }
